+++ doc/changes.rst
@@ -4,4 +4,8 @@ Changes
 Since 0.0.4
 -----------
 
-Nothing yet
+* :class:`~eelbrain.vessels.experiment.mne_experiment`
+
+	* Ability to use manual epoch rejection instead of the automatic procedure.
+	  The `epoch_rejection` class attribute specifies how epochs should be 
+	  rejected.